Grid block converted
Conversion of grid complete.

Grid can support a more robust min() value, such as a calc(var(--measure) / 3) function, in the future we need to think of a more robust way to handle entry of values like that, versus simple static values, although those will suffice for now.

In a 0.3.0 run through, we'll have to think about a more standard way of normalizing sizing and spacing CSS classes and their corresponding values.

Another thought is to remove the namespace prefix from each var in a block, e.g. --wf-box--space, and simply go with something like --wf--space and see how that affects the inheritance. Perhaps a toggle between "don't inherit" and inherit.