Major clean up and re-organization
Can almost eliminate the old 'editor-components' dir in block-library. The rest have been migrated to block-editor dir in the master area.

This is so that we can package up block-editor as an NPM package for more complex blocks or other plugin sets, as well as keeping the block-library code more clean and concise.

Almost finished with refactor of imposter.